code/requirements.txt
allure-pytest==2.8.32
allure-python-commons==2.8.32
apipkg==1.5
Appium-Python-Client==1.1.0
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
execnet==1.8.0
Faker==8.0.0
Flask==1.1.2
greenlet==1.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
jenkins-jobs==0.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multi-key-dict==2.0.3
packaging==20.9
pbr==5.6.0
pluggy==0.13.1
py==1.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.1
pytest-forked==1.3.0
pytest-xdist==2.2.1
python-dateutil==2.8.1
python-jenkins==1.7.0
requests==2.25.1
selenium==3.141.0
six==1.16.0
SQLAlchemy==1.4.13
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
webdriver-manager==3.3.0
Werkzeug==2.0.1
xmltodict==0.12.0
